# src/frontend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Fitness Tracking API
  version: 1.0.0
  description: |
    A comprehensive fitness tracking API with voice integration, workout management, 
    exercise libraries, and LLM-powered workout generation capabilities.
  
    ## Features
    - User authentication with JWT tokens
    - Workout creation and tracking
    - Exercise library with media support
    - Voice-powered workout logging
    - LLM workout generation
    - Goal setting and progress tracking
    - Offline support with sync capabilities
    
  contact:
    name: API Support
    email: support@fitnessapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.fitnessapp.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /workouts:
    get:
      tags:
        - Workouts
      summary: Get workouts
      description: Retrieve user's workouts with pagination and filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkoutStatus'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Workouts retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedWorkouts'

    post:
      tags:
        - Workouts
      summary: Create workout
      description: Create a new workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutCreateRequest'
      responses:
        '201':
          description: Workout created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workout'
  /workouts/{id}:
    get:
      tags:
        - Workouts
      summary: Get workout by ID
      description: Retrieve a specific workout by its ID
      parameters:
        - $ref: '#/components/parameters/WorkoutId'
      responses:
        '200':
          description: Workout retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workout'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Workouts
      summary: Update workout
      description: Update an existing workout
      parameters:
        - $ref: '#/components/parameters/WorkoutId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutUpdateRequest'
      responses:
        '200':
          description: Workout updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workout'

    delete:
      tags:
        - Workouts
      summary: Delete workout
      description: Delete a workout
      parameters:
        - $ref: '#/components/parameters/WorkoutId'
      responses:
        '200':
          description: Workout deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /workouts/generate:
    post:
      tags:
        - Workouts
        - LLM Integration
      summary: Generate AI workout
      description: Generate a workout using LLM based on user prompt and preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutGenerateRequest'
      responses:
        '201':
          description: Workout generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Workout'
        '400':
          $ref: '#/components/responses/BadRequestError'

  # Exercise Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    WorkoutId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Workout ID
      example: "workout_123"
...
  schemas:
...
    WorkoutSet:
      type: object
      required:
        - id
        - exerciseId
        - setNumber
        - completed
      properties:
        id:
          type: string
          example: "set_123"
        exerciseId:
          type: string
          example: "exercise_bench_press"
        setNumber:
          type: integer
          minimum: 1
          example: 1
        reps:
          type: integer
          minimum: 0
          example: 8
        weight:
          type: number
          minimum: 0
          example: 135.5
        time:
          type: integer
          minimum: 0
          description: Duration in seconds
          example: 30
        distance:
          type: number
          minimum: 0
          description: Distance in meters
          example: 1000
        rpe:
          type: integer
          minimum: 1
          maximum: 10
          description: Rate of Perceived Exertion (1-10)
          example: 7
        notes:
          type: string
          example: "Felt good, could do more"
        completed:
          type: boolean
          example: true
        completedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    Workout:
      type: object
      required:
        - id
        - userId
        - name
        - date
        - sets
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: "workout_123"
        userId:
          type: string
          example: "user_123"
        name:
          type: string
          example: "Push Day"
        description:
          type: string
          example: "Chest, shoulders, and triceps workout"
        date:
          type: string
          format: date
          example: "2024-01-15"
        startTime:
          type: string
          format: date-time
          example: "2024-01-15T09:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        sets:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSet'
        status:
          $ref: '#/components/schemas/WorkoutStatus'
        totalVolume:
          type: number
          description: Total volume in kg/lbs
          example: 5400
        totalTime:
          type: integer
          description: Total workout time in seconds
          example: 5400
        notes:
          type: string
          example: "Great workout, felt strong"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...
    MuscleGroup:
      type: string
      enum:
        - chest
        - back
        - shoulders
        - biceps
        - triceps
        - forearms
        - core
        - glutes
        - quadriceps
        - hamstrings
        - calves
        - traps
        - lats
      example: chest
    Equipment:
      type: string
      enum:
        - barbell
        - dumbbell
        - kettlebell
        - cable
        - machine
        - bodyweight
        - resistance_band
        - medicine_ball
        - suspension_trainer
        - none
      example: barbell
    WorkoutStatus:
      type: string
      enum:
        - planned
        - in_progress
        - completed
        - cancelled
      example: completed
...
    WorkoutCreateRequest:
      type: object
      required:
        - name
        - date
      properties:
        name:
          type: string
          example: "Push Day"
        description:
          type: string
          example: "Chest, shoulders, and triceps workout"
        date:
          type: string
          format: date
          example: "2024-01-15"
        exerciseIds:
          type: array
          items:
            type: string
          example: ["exercise_bench_press", "exercise_shoulder_press"]
    WorkoutUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: "Updated Push Day"
        description:
          type: string
        status:
          $ref: '#/components/schemas/WorkoutStatus'
        notes:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
...
    WorkoutGenerateRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          example: "Create a 45-minute upper body strength workout for intermediate level"
        preferences:
          type: object
          properties:
            duration:
              type: integer
              description: Desired workout duration in minutes
              example: 45
            equipment:
              type: array
              items:
                $ref: '#/components/schemas/Equipment'
              example: ["barbell", "dumbbell"]
            targetMuscles:
              type: array
              items:
                $ref: '#/components/schemas/MuscleGroup'
              example: ["chest", "shoulders", "triceps"]
            level:
              type: string
              enum: ["beginner", "intermediate", "advanced"]
              example: "intermediate"
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
...
    PaginatedResponse:
      type: object
      required:
        - data
        - total
        - page
        - limit
        - hasNext
        - hasPrev
      properties:
        data:
          type: array
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    PaginatedWorkouts:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Workout'
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: object
              example: {"field": "email", "issue": "Invalid email format"}
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "BAD_REQUEST"
              message: "Invalid request data"
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
...